Grid attempt 2
Added upscale button and drop down menus for downloads and upscales. Only first view has the menus due to weirdness with the way views are handled. Added /batch command to retrieve batch images beyond the first 25 with a batch_id and image_id. Corrected issues with batches that are multiples of 25 resulting in a divide by zero error.

Remove stealth png